# backend/src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "API de Cursos",
      "version": "1.0.0",
      "description": "API para administrar cursos"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
      "/cursos": {
        "get": {
          "tags": ["Cursos"],
          "summary": "Obtener todos los cursos",
          "responses": {
            "200": {
              "description": "Lista de cursos",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Curso"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        },
        "post": {
          "tags": ["Cursos"],
          "summary": "Crear un nuevo curso",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Curso"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Curso creado satisfactoriamente",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Curso"
                  }
                }
              }
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        }
      },
      "/cursos/{id}": {
        "get": {
          "tags": ["Cursos"],
          "summary": "Obtener un curso por su ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID del curso",
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Curso encontrado",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Curso"
                  }
                }
              }
            },
            "404": {
              "description": "Curso no encontrado"
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        }
               
    }
}
}